fix(client): address Devin review feedback on CLI agent control v1
Shutdown correctness — controlserver no longer detaches connection
threads. Stop() now closes client sockets to unblock recv(), drains
queued + pendingWaits replies via a callback registered by Game::Load,
then joins all handler threads. Fixes UAF where detached handlers
read the destroyed `running` atomic, and fixes the multi-frame wait
hang where wait_frames/wait_ms/wait_for_state/step left handler
threads blocked on fut.get() forever on shutdown.

Defensive cleanups:
- FindWidgetByLabel mask widened to Uint64; some object types are
  >= 32 (WALLPROJECTILE, BASEEXIT) and 1u<<type is UB there.
- HandleSDLEvents early-returns in headless mode (would deref NULL
  window if any event ever fired without SDL_INIT_VIDEO).
- DrainControlQueue fails POST_RENDER ops in dedicated server mode
  (PostFrameReplies isn't called there, would hang screenshots).
- select handler distinguishes WIDGET_AMBIGUOUS from WIDGET_NOT_FOUND,
  matching click handler.
- Drop dead `stepFramesRemaining < 0` branch (int, never negative).
- Comment headless palette ordering near SDL init.

Build fix: shared.h dropped legacy `#define snprintf _snprintf`
(pre-VS2015 compat) — it mangled `(std::snprintf)` inside vendored
nlohmann/json on modern MSVC.

+++ clients/silencer/src/interface.cpp
@@ -705,14 +705,14 @@ static const char* LabelOf(Object* o){
 }
 
 Interface::WidgetMatch Interface::FindWidgetByLabel(World& world,
-	const char* labelOrId, Uint32 wantedTypes, Uint16* outId) const {
+	const char* labelOrId, Uint64 wantedTypes, Uint16* outId) const {
 	if(!labelOrId || !*labelOrId) return MATCH_NOT_FOUND;
 	// Numeric ID path: caller passed a literal Uint16.
 	char* endp = nullptr;
 	long asnum = std::strtol(labelOrId, &endp, 10);
 	if(endp && *endp == 0 && asnum > 0 && asnum <= 0xFFFF){
 		Object* o = world.GetObjectFromId((Uint16)asnum);
-		if(o && (wantedTypes == 0 || (wantedTypes & (1u << o->type)))){
+		if(o && (wantedTypes == 0 || (wantedTypes & (1ULL << o->type)))){
 			*outId = (Uint16)asnum;
 			return MATCH_OK;
 		}
@@ -723,7 +723,7 @@ Interface::WidgetMatch Interface::FindWidgetByLabel(World& world,
 	for(Uint16 oid : objects){
 		Object* o = world.GetObjectFromId(oid);
 		if(!o) continue;
-		if(wantedTypes != 0 && !(wantedTypes & (1u << o->type))) continue;
+		if(wantedTypes != 0 && !(wantedTypes & (1ULL << o->type))) continue;
 		const char* label = LabelOf(o);
 		if(label && IEq(label, labelOrId)){
 			if(hits == 0) firstHit = oid;
